refactor: rework pull_request.opened event handler
