+++ src/config.ts
@@ -1,11 +1,15 @@
 import { HandlerModule } from "./common";
 import StarCreateHandler from "./events/star.created";
+import PullRequestOpenHandler from "./events/pull_request.opened";
 
 export interface Configuration {
   app_name: string;
 }
 
-export const Handlers: HandlerModule[] = [StarCreateHandler];
+export const Handlers: HandlerModule[] = [
+  StarCreateHandler,
+  PullRequestOpenHandler,
+];
 
 export const AppConfig: Configuration = {
   app_name: process.env.APP_NAME!,
