+++ src/index.ts
@@ -372,114 +372,10 @@ ${context.payload.issue.body!.split("<!-- BEGIN CHANGELOGS -->")[1]}
   });
 
   // on receive pull_request.opened event
-  app.on(
-    "pull_request.opened",
-    async (context: Context<"pull_request.opened">) => {
-      const metadata = {
-        repo: context.payload.repository.name,
-        owner: context.payload.organization?.login as string,
-        default_branch: context.payload.repository.default_branch,
-        html_url: context.payload.repository.html_url,
-        pull_request: {
-          ref: context.payload.pull_request.head.ref,
-          title: context.payload.pull_request.title,
-          author: context.payload.pull_request.user.login,
-          number: context.payload.pull_request.number,
-          updated_at: context.payload.pull_request.updated_at,
-          html_url: context.payload.pull_request.html_url,
-        },
-      };
-
-      app.log.info(
-        `received a pull_request.opened event: ${JSON.stringify(metadata)}`
-      );
-
-      // case_#1: automatically assign assignee if not present
-      // 1.1 assign pull_request author to be the default assignee
-      // https://octokit.github.io/rest.js/v18#issues-add-assignees
-      const author = metadata.pull_request.author.includes("bot")
-        ? "daebot"
-        : metadata.pull_request.author;
-      await context.octokit.issues.addAssignees({
-        owner: metadata.owner,
-        repo: metadata.repo,
-        issue_number: metadata.pull_request.number,
-        assignees: [author],
-      });
-
-      // 1.2 audit event
-      const msg = `👷 PR - [#${metadata.pull_request.number}](${metadata.pull_request.html_url}) is raised in ${metadata.repo}; assign @${author} as the default assignee`;
-
-      app.log.info(msg);
-
-      const tg = new TelegramClient();
-      await tg.sendMsg(msg, [
-        process.env.TELEGRAM_DAEUNIVERSE_AUDIT_CHANNEL_ID!,
-      ]);
-
-      // case_#2: automatically assign label if not present, default label should align with "kind" as part of the pr title
-
-      // 1.1 automatically add label(s) to pull_request
-      const defaultLables = [
-        "fix",
-        "feat",
-        "feature",
-        "patch",
-        "ci",
-        "optimize",
-        "chore",
-        "refactor",
-        "style",
-        "doc",
-        "docs",
-      ];
-
-      // https://octokit.github.io/rest.js/v18#issues-list-labels-on-issue
-      const prOpenedLabels = await context.octokit.issues
-        .listLabelsOnIssue({
-          owner: metadata.owner,
-          repo: metadata.repo,
-          issue_number: metadata.pull_request.number,
-        })
-        .then((res) => res.data);
-
-      if (prOpenedLabels.length == 0) {
-        const labels = defaultLables
-          .filter((label: string) =>
-            metadata.pull_request.title.startsWith(label)
-          )
-          .map((item) => {
-            if (item == "feat") item = "feature";
-            if (item == "docs" || item == "doc") item = "documentation";
-            return item;
-          });
-
-        if (labels.length > 0) {
-          const msg = `🏷 PR - [#${metadata.pull_request.number}](${
-            metadata.pull_request.html_url
-          }) in ${metadata.repo} is missing labels; added ${JSON.stringify(
-            labels
-          )}`;
-
-          // https://octokit.github.io/rest.js/v18#issues-add-labels
-          await context.octokit.issues.addLabels({
-            owner: metadata.owner,
-            repo: metadata.repo,
-            issue_number: metadata.pull_request.number,
-            labels: labels,
-          });
-
-          // 1.2 audit event
-          app.log.info(msg);
-
-          const tg = new TelegramClient();
-          await tg.sendMsg(msg, [
-            process.env.TELEGRAM_DAEUNIVERSE_AUDIT_CHANNEL_ID!,
-          ]);
-        }
-      }
-    }
-  );
+  app.on("pull_request.opened", async (context: Context<any>) => {
+    const result = await Run(context, app, "pull_request.opened");
+    app.log.info(JSON.stringify(result));
+  });
 
   // on pull_request.synchronize event
   app.on(
